feat(runningway): 启用 fp16 计算并更新模型配置与训练逻辑
- 将 CUDA 内核中的 bfloat16 支持替换为 fp16，提升兼容性与性能
- 更新 config.json 中的默认加载模型路径
- 调整模型代码中环境变量 RNW_MY_TESTING 默认值为 r010
- 优化 DeepSpeed 优化器参数命名及初始化逻辑
- 修复梯度检查点包装函数以适配非张量输入限制
- 增强状态池权重和缓冲区保存逻辑
- 完善模型加载判断条件，确保文件存在且有效
- 更新精度设置格式以匹配 PyTorch Lightning 1.9.5 要求
- 微调日志输出图标与描述文本
- 添加 NCCL 相关环境变量以增强分布式训练稳定性

+++ 1.0/main/cuda/wkv7_op.cpp
@@ -1,6 +1,6 @@
 #include <torch/extension.h>
-#include <cuda_bf16.h>
-using bf = __nv_bfloat16;
+#include <cuda_fp16.h> // 使用 fp16
+using bf = __half; // 将原来的 __nv_bfloat16 改为 __half
 
 void cuda_forward(int B, int T, int H, bf*w, bf*q, bf*k, bf*v, bf*z, bf*a, bf*y, float*s, float*sa);
